+++ src/main/groovy/com/electro/bikeapp/dtos/ShiftsDTO.groovy
@@ -10,6 +10,6 @@ class ShiftsDTO {
 
     String exitTimes
 
-    BigDecimal totalWeeklyHours
+    Float totalWeeklyHours
 
 }
