+++ src/main/groovy/com/electro/bikeapp/repositories/InventoryRepository.groovy
@@ -20,8 +20,8 @@ interface InventoryRepository extends JpaRepository<BikeDomain, Long> {
     @Query('SELECT b FROM BikeDomain b WHERE (:bikeColor is null or b.color = :bikeColor) and (:priceMin is null or b.retailCost >= :priceMin) and (:priceMax is null or b.retailCost <= :priceMax) and (:inStock is null or b.inStock = :inStock) and (:condition is null or b.condition = :condition) and (:make is null or b.make = :make)')
     // This method returns a list of bikes. It accepts the name, make, color, lessThanCost, and greaterThanCost
     List<BikeDomain> allParamSearch(@Param('bikeColor') String bikeColor,
-                                    @Param('priceMin') BigDecimal priceMin,
-                                    @Param('priceMax') BigDecimal priceMax,
+                                    @Param('priceMin') Float priceMin,
+                                    @Param('priceMax') Float priceMax,
                                     @Param('inStock') Boolean inStock,
                                     @Param('condition') String condition,
                                     @Param('make') String make);
